+++ NSTree/NSTree.m
@@ -149,9 +149,9 @@
 
 @interface NSTree()
     @property (nonatomic, strong) NSTreeNode *root;
-    @property (nonatomic, assign) int nodeMinimum;
-    @property (nonatomic, assign, readwrite) int nodeCapacity; 
-    @property (nonatomic, assign, readwrite) int count;
+    @property (nonatomic, assign) NSUInteger nodeMinimum;
+    @property (nonatomic, assign, readwrite) NSUInteger nodeCapacity;
+    @property (nonatomic, assign, readwrite) NSUInteger count;
     
     // Cache and flag for quick access
     @property (nonatomic, assign, readwrite) bool cacheOutdated; 
@@ -181,7 +181,7 @@
 }
 
 /** @brief Create tree with a certain number of allowable children */
-- (id)initWithNodeCapacity:(int)nodeCapacity
+- (id)initWithNodeCapacity:(NSUInteger)nodeCapacity
 {
     self = [super init];
     if (self) {
@@ -195,7 +195,7 @@
 }
 
 /** @brief Create tree with a certain number of allowable children using the given array of objects as its base data */
-- (id)initWithNodeCapacity:(int)nodeCapacity withSortedObjects:(NSArray *)data
+- (id)initWithNodeCapacity:(NSUInteger)nodeCapacity withSortedObjects:(NSArray *)data
 {
     self = [super init];
     if (self) {
@@ -219,7 +219,7 @@
     @param data NSArray of objects, must be sorted. 
     @return NSTreeNode * Root of data tree.
 */
-- (NSTreeNode *)buildTreeWithNodeCapacity:(int)nodeCapacity withSortedObjects:(NSArray *)data
+- (NSTreeNode *)buildTreeWithNodeCapacity:(NSUInteger)nodeCapacity withSortedObjects:(NSArray *)data
 {
     NSMutableArray *children = [NSMutableArray new];
     NSMutableArray *parents = [NSMutableArray new]; 
@@ -237,9 +237,9 @@
         prev = child;   // update previous
         
         // Fill it with max capacity + 1 data, except for last node
-        int fillCount = nodeCapacity 
+        NSUInteger fillCount = nodeCapacity
             + (data.count - i > nodeCapacity + 1 ? 1 : 0);
-        for (int j = 0; j < fillCount && i < data.count; ++j, ++i) {
+        for (NSUInteger j = 0; j < fillCount && i < data.count; ++j, ++i) {
             [child.data addObject:data[i]];
         }
         
@@ -268,12 +268,12 @@
             prev = parent;  // update previous
             
             // Fill it with data & children
-            int fillCount = nodeCapacity 
+            NSUInteger fillCount = nodeCapacity
                 + (children.count - i > nodeCapacity + 1 ? 1 : 0); 
-            for (int j = 0; j < fillCount && i < children.count - 1; ++j, ++i) 
+            for (NSUInteger j = 0; j < fillCount && i < children.count - 1; ++j, ++i)
             {
                 child = children[i];  
-                int index = child.data.count - 1;
+                NSUInteger index = child.data.count - 1;
                                
                 // Add child
                 [parent.children addObject:child];
@@ -455,7 +455,7 @@
 }
 
 /** @brief Returns object at index, or nil if none / out of bounds */
-- (id)objectAtIndex:(int)index
+- (id)objectAtIndex:(NSInteger)index
 {
     // Insanity checks
     if (index < 0) {
@@ -504,7 +504,7 @@
     }
    
     // Find index where we should put it, and add it
-    int index = [node.data indexOfObject:object 
+    NSUInteger index = [node.data indexOfObject:object
         inSortedRange:NSMakeRange(0, node.data.count) 
         options:NSBinarySearchingInsertionIndex 
         usingComparator:^NSComparisonResult(id obj1, id obj2) {
@@ -559,7 +559,7 @@
 //    NSLog(@"Removing object %@ from node %@", object, node);
     
     // Get index to remove from
-    int index = [node indexOfDataObject:object];
+    NSUInteger index = [node indexOfDataObject:object];
     if (index == NSNotFound) {
         NSLog(@"Warning! Could not find index of object for removal: %@", object);
         return false;
@@ -597,7 +597,7 @@
     }
     
     // Search for item in node data
-    int index = [node.data indexOfObject:object 
+    NSInteger index = [node.data indexOfObject:object
         inSortedRange:NSMakeRange(0, node.data.count) 
         options:NSBinarySearchingInsertionIndex 
             | NSBinarySearchingFirstEqual
@@ -632,7 +632,7 @@
 //    NSLog(@"Get: %@ in %@", object, node);
     
     // Search for item in node data
-    int index = [node.data indexOfObject:object 
+    NSInteger index = [node.data indexOfObject:object
         inSortedRange:NSMakeRange(0, node.data.count) 
         options:NSBinarySearchingInsertionIndex 
             | NSBinarySearchingFirstEqual
@@ -671,7 +671,7 @@
     if (node.children.count)
     {
         // Search for item in node data
-        int index = [node.data indexOfObject:object 
+        NSInteger index = [node.data indexOfObject:object
             inSortedRange:NSMakeRange(0, node.data.count) 
             options:NSBinarySearchingInsertionIndex 
                 | NSBinarySearchingFirstEqual 
@@ -816,15 +816,15 @@
 
         // Create right node to be efficient about removing from arrays
         NSTreeNode *newRightNode = [[NSTreeNode alloc] initWithParent:node.parent];
-        int middle = node.data.count / 2;
-        int childIndex = middle + 1;
+        NSUInteger middle = node.data.count / 2;
+        NSUInteger childIndex = middle + 1;
         id object = node.data[middle];
 
         // Iterate through data & children from middle + 1 and add to new node
-        for (int i = childIndex; i < node.data.count; ++i) {
+        for (NSUInteger i = childIndex; i < node.data.count; ++i) {
             [newRightNode.data addObject:node.data[i]];
         }
-        for (int i = childIndex; i < node.children.count; ++i) {
+        for (NSUInteger i = childIndex; i < node.children.count; ++i) {
             [newRightNode.children addObject:node.children[i]];
             [node.children[i] setParent:newRightNode]; 
         } 
@@ -913,21 +913,21 @@
     }
     
     // Get index of node in children array of parent
-    int indexOfChild = [node.parent indexOfChildNode:node];
+    NSUInteger indexOfChild = [node.parent indexOfChildNode:node];
     if (indexOfChild == NSNotFound) {
         NSLog(@"Warning! Could not find index of child in parent: %@", node);
         return;
     }
     
     // Insert parent data that is next to the node
-    int indexOfParentData = indexOfChild - direction;
-    int indexOfInsert = (direction ? 0 : node.data.count);
+    NSUInteger indexOfParentData = indexOfChild - direction;
+    NSUInteger indexOfInsert = (direction ? 0 : node.data.count);
     [node.data insertObject:node.parent.data[indexOfParentData] 
                     atIndex:indexOfInsert];
     
     // Replace parent data with data from sibling
     NSTreeNode *sibling = (direction ? node.previous : node.next);
-    int indexOfRemove = (direction ? sibling.data.count - 1 : 0); 
+    NSUInteger indexOfRemove = (direction ? sibling.data.count - 1 : 0);
     [node.parent.data replaceObjectAtIndex:indexOfParentData 
                                 withObject:sibling.data[indexOfRemove]];
     [sibling.data removeObjectAtIndex:indexOfRemove];
@@ -983,7 +983,7 @@
     
     // Find index of separator object in parent
     parent = leftNode.parent;
-    int index = [parent indexOfChildNode:leftNode];
+    NSUInteger index = [parent indexOfChildNode:leftNode];
     
     // Transfer data & children over from parent / right node
     [leftNode.data addObject:parent.data[index]];
